+++ libs/providers/flagd-web/src/lib/flagd-web-provider.spec.ts
@@ -10,8 +10,8 @@ import {
   StandardResolutionReasons,
 } from '@openfeature/web-sdk';
 import fetchMock from 'jest-fetch-mock';
-import { Service } from '../proto/ts/schema/v1/schema_connect';
-import { AnyFlag, EventStreamResponse, ResolveAllResponse } from '../proto/ts/schema/v1/schema_pb';
+import { Service } from '../proto/ts/flagd/evaluation/v1/evaluation_connect';
+import { AnyFlag, EventStreamResponse, ResolveAllResponse } from '../proto/ts/flagd/evaluation/v1/evaluation_pb';
 import { FlagdWebProvider } from './flagd-web-provider';
 
 const EVENT_CONFIGURATION_CHANGE = 'configuration_change';