+++ libs/providers/flagd-web/src/lib/flagd-web-provider.ts
@@ -16,8 +16,8 @@ import {
   StandardResolutionReasons,
   TypeMismatchError,
 } from '@openfeature/web-sdk';
-import { Service } from '../proto/ts/schema/v1/schema_connect';
-import { AnyFlag } from '../proto/ts/schema/v1/schema_pb';
+import { Service } from '../proto/ts/flagd/evaluation/v1/evaluation_connect';
+import { AnyFlag } from '../proto/ts/flagd/evaluation/v1/evaluation_pb';
 import { FlagdProviderOptions, getOptions } from './options';
 
 export const ERROR_DISABLED = 'DISABLED';