+++ libs/providers/flagd/src/e2e/constants.ts
@@ -1,0 +1,4 @@
+export const FLAGD_NAME = 'flagd Provider';
+export const E2E_CLIENT_NAME = 'e2e';
+export const UNSTABLE_CLIENT_NAME = 'unstable';
+export const UNAVAILABLE_CLIENT_NAME = 'unavailable';