+++ libs/shared/flagd-core/src/lib/flagd-core.ts
@@ -137,7 +137,7 @@ export class FlagdCore implements Storage {
   resolve<T extends FlagValue>(
     type: FlagValueType,
     flagKey: string,
-    defaultValue: T,
+    _: T,
     evalCtx: EvaluationContext = {},
     logger?: Logger,
   ): ResolutionDetails<T> {
@@ -150,11 +150,7 @@ export class FlagdCore implements Storage {
 
     // flag status check
     if (flag.state === 'DISABLED') {
-      logger.debug(`Flag ${flagKey} is disabled, returning default value ${defaultValue}`);
-      return {
-        value: defaultValue,
-        reason: StandardResolutionReasons.DISABLED,
-      };
+      throw new FlagNotFoundError(`flag: '${flagKey}' is disabled`);
     }
 
     let variant;
@@ -169,11 +165,11 @@ export class FlagdCore implements Storage {
       try {
         targetingResolution = this._targeting.applyTargeting(flagKey, flag.targeting, evalCtx);
       } catch (e) {
-        logger.error(`Error evaluating targeting rule for flag ${flagKey}, falling back to default`, e);
-        targetingResolution = null;
+        throw new GeneralError(`Error evaluating targeting rule for flag ${flagKey}: ${(e as Error)?.message}`);
       }
 
-      if (!targetingResolution) {
+      // Return default variant if targeting resolution is null or undefined
+      if (targetingResolution == null) {
         variant = flag.defaultVariant;
         reason = StandardResolutionReasons.DEFAULT;
       } else {